associate_phone_number_to_persons: remove \n in JSON query to phone service; better logging

+++ associate_phone_number_to_persons/run.sh
@@ -10,13 +10,15 @@ echo idlist.txt populated;   wc idlist.txt
 mkdir $RESULTS
 PHONE_JSON=/tmp/output.json
 
+NUM=0
 for ID in `cat idlist.txt`	# "EVTC001100001" # "EVTC075140343"
 do
+  NUM=`expr $NUM + 1`
   echo "For person $ID"
   # remove end of line:
   ID=${ID:0:`expr length $ID - 1`}
   if test $ID = "ID" ; then continue ; fi
-  echo "For person <$ID>"
+  echo "For person <$ID> , n° $NUM - `date`"
   RESULT=$RESULTS/$ID.ttl
   RESULTJSON=$RESULTS/$ID.json
   if test -f $RESULTJSON ; then echo $ID already done; continue; fi
@@ -26,15 +28,17 @@ do
   $SPARQL_COMMAND --query person_to_phone_query.rq > /tmp/phone_query.ttl
   turtle --output=JSON-LD /tmp/phone_query.ttl >  /tmp/phone_query.json
   echo before index.js
+  PHONE_QUERY=`cat /tmp/phone_query.json | tr '\n' ' ' | sed 's/\\\n//g' `
+  echo PHONE_QUERY "$PHONE_QUERY"
   rm -f $PHONE_JSON
-  nodejs ../../semantize/fetch/vtc/phones/index.js "`cat /tmp/phone_query.json | tr '\n' ' ' `"
+  nodejs ../../semantize/fetch/vtc/phones/index.js "$PHONE_QUERY"
   RESULT_LENGTH=`wc -c < $PHONE_JSON`
   if test $RESULT_LENGTH -lt 10 ; then continue ; fi
   cp $PHONE_JSON $RESULTJSON
 
   # launch Silk
   java -cp $JENAPATH riotcmd.riot --syntax=JSON-LD --output=turtle $PHONE_JSON > /tmp/output.ttl
-  cp /tmp/output.ttl $RESULT
+  if test -s  /tmp/output.ttl ; then cp /tmp/output.ttl $RESULT ; fi
   echo "prefix oxi: <http://omerxi.com/ontologies/core.owl.ttl#>  CONSTRUCT { ?S ?P ?O . } WHERE { ?S oxi:id \"$ID\" ; ?P ?O . }" > /tmp/extractbyid.rq
   $SPARQL_COMMAND --query /tmp/extractbyid.rq --results TURTLE > /tmp/extractbyid.ttl
   java -jar -DconfigFile=../silk/foaf_name.vcard_street-adress/configFile.xml -DlogQueries=false $SILK/silk.jar
